libsemanage: add support for HLL to CIL compilers
An HLL to CIL compiler must exist in the compiler_directory
path which is configubrable in semanage.conf. By default, this path
is /usr/libexec/selinux/hll/. The compiler name needs to match the HLL
language extension. For example, for pp files, /usr/libexec/selinux/hll/pp must exist.

The HLL module is then bunzip2'd, piped to the HLL to CIL compiler, and a CIL module is created.
A the HLL and CIL files are bzip2'd and placed in <store_path>/tmp/modules/<priority>/<module>/

Signed-off-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>