libselinux: modify load_policy to be more forgiving of libsepol versions
At one point, load_policy.c dlopen'ed libsepol.so, but there were cases
where the .so didn't exist, so it was modified to dlopen libsepol.so.1
instead (selinux-historical, b35015295bf1e2aaea1e0cfef1e301726d0142a9),
which did always exist. However, with the .so bump of libsepol to
integrate CIL and support source policies, this is now referencing the
wrong library. To be a little less error prone, first try to dlopen
libsepol.so, if that doesn't exist, then try to dlopen specific versions
of libsepol.

Signed-off-by: Steve Lawrence <slawrence@tresys.com>